_Dockerized/linkedin_analytics/web/requirements.txt
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
dj-database-url==0.5.0
Django==2.1.7
django-forms-bootstrap==3.1.0
django-materializecss-form==1.1.11
django-stubs==0.9.0
djangorestframework==3.9.1
djangorestframework-stubs==0.2.0
idna==2.8
mypy==0.670
mypy-extensions==0.4.1
numpy==1.16.2
pandas==0.24.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
typed-ast==1.3.1
typing-extensions==3.7.2
urllib3==1.24.1
